requirements.txt
flask
flask-cors
gunicorn
torch
torchvision
pillow
transformers
opencv-python-headless
numpy